Add documentation for AnimationTimer
